requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26.4  # Compatible with PyTorch 2.2
opencv-python>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
click>=8.1.0

# AI upscaling models
basicsr==1.4.2
realesrgan==0.3.0
gfpgan>=1.3.8
facexlib>=0.3.0

# Display utilities
colorama>=0.4.6
humanize>=4.9.0

# Model management
gdown>=4.7.0